+++ pagination/compose/src/jvmTest/kotlin/InfinityPagedComponentTests.kt
@@ -30,9 +30,7 @@ class InfinityPagedComponentTests {
                 }
             ) {
                 if (it == null) {
-                    if (this.iterationState.value.second != null) {
-                        assertEquals(0, (this.iterationState.value.second as? SimplePagination) ?.page)
-                    }
+                    assertEquals(0, this.currentlyLoadingPage.value ?.page)
                 } else {
                     assertEquals(expectedList, it)
                 }