+++ CHANGELOG.md
@@ -6,6 +6,15 @@
   * `Cache`:
     * All cache repos now do not have `open` vals - to avoid collisions in runtime
 
+## 0.24.9
+
+* `Pagination`:
+  * Make alternative constructor parameter `size` of `PaginationResult` with default value
+  * Add `Pagination.previousPage` extension
+  * `Compose`:
+    * Rework of `InfinityPagedComponentContext`
+    * Rework of `PagedComponent`
+
 ## 0.24.8
 
 * `Versions`: